refactor(config): move nuxt config js to folder
