+++ nuxt-config/vuetify-icon.js
@@ -1,0 +1,11 @@
+module.exports = {
+  moon: 'far fa-moon',
+  sun: 'far fa-sun',
+  version: 'fas fa-at',
+  home: 'fas fa-home',
+  beta: 'fas fa-flask',
+  admin: 'fas fa-user-shield',
+  github: 'fab fa-github',
+  docs: 'fas fa-book',
+  apis: 'fas fa-code'
+}
