+++ nuxt-config/vuetify-theme.js
@@ -1,0 +1,11 @@
+module.exports = {
+  primary: '#60C1E8',
+  secondary: '#FFA6C9',
+  accent: '#B388FF', // not change
+  error: '#FF6666',
+  info: '#2196F3', // not change
+  success: '#96DFA9',
+  warning: '#FFC107', // not change
+  favnet: '#FF3333',
+  favprang: '#FFE97F'
+}
