+++ nuxt.config.js
@@ -1,58 +1,23 @@
 const path = require('path')
 const serveStatic = require('serve-static')
 
-const pkg = require('./package')
+const env = process.env.NODE_ENV
 
 const name = 'KcNt Portfolio'
-const env = process.env.NODE_ENV
-const branch = process.env.BRANCH
+const pkg = require('./package')
 
-const onesignal_dev = process.env.ONESIGNAL_DEV_APPID
-const onesignal_prod = process.env.ONESIGNAL_APPID
-
-const isProd = env === 'production'
-const isDev = env === 'development'
-
-const baseUrl = isProd ? 'https://kcnt.info' : 'http://localhost:3000'
-
-const generateIconPath = () => {
-  const size = ['dot5x', '1x', '300ppi']
-  const type = ['round-icon', 'icon']
-  const color = ['dark', 'light', 'primary']
-
-  const result = []
-
-  color.forEach(c => {
-    size.forEach(s => {
-      type.forEach(t => {
-        const afterIconPath = s === 'dot5x' ? '0.5x' : s
-        const sizePath = s === '1x' ? '' : `-${s}`
-        const postfix = s === '300ppi' ? '-high' : s === 'dot5x' ? '@0.5x' : ''
-
-        result.push({
-          from: `/logo/png/${c}${t
-            .replace(/[-]?icon/g, '')
-            .replace('round', '-round')}${sizePath}`,
-          to: `/resources/images/icon/${afterIconPath}/${c}-${t}${postfix}.png`
-        })
-      })
-    })
-  })
-
-  color.forEach(c => {
-    type.forEach(t => {
-      result.push({
-        from: `/logo/svg/${c}${t
-          .replace(/[-]?icon/g, '')
-          .replace('round', '-round')}`,
-        to: `/resources/images/icon/svg/${c}-${t}.svg`
-      })
-    })
-  })
-
-  // console.log(result.map(v => v.from)) // path all icon redirect path
-
-  return result
+const data = {
+  pkg,
+  name,
+  env,
+  branch: process.env.BRANCH,
+  baseUrl: env === 'production' ? 'https://kcnt.info' : 'http://localhost:3000',
+  isProd: env === 'production',
+  isDev: env === 'development',
+  onesignal: {
+    devID: process.env.ONESIGNAL_DEV_APPID,
+    prodID: process.env.ONESIGNAL_APPID
+  }
 }
 
 module.exports = {
@@ -62,35 +27,7 @@ module.exports = {
    */
   head: {
     title: name,
-    meta: [
-      {
-        charset: 'utf-8'
-      },
-      {
-        name: 'viewport',
-        content: 'width=device-width, initial-scale=1'
-      },
-      {
-        hid: 'description',
-        name: 'description',
-        content: pkg.description
-      },
-      {
-        hid: 'keywords',
-        name: 'keywords',
-        content: 'Portfolio,VueJS,JS,CSS,Website,Nuxt,Personal'
-      },
-      {
-        hid: 'author',
-        name: 'author',
-        content: pkg.author
-      },
-      {
-        hid: 'version',
-        name: 'version',
-        content: pkg.version
-      }
-    ],
+    meta: require('./nuxt-config/meta')(data),
     script: [
       {
         innerHTML:
@@ -114,7 +51,7 @@ module.exports = {
   ],
 
   env: {
-    baseUrl: baseUrl,
+    baseUrl: data.baseUrl,
     version: pkg.version,
     license: pkg.license,
     buildDate: +new Date(),
@@ -165,57 +102,9 @@ module.exports = {
       }
     ],
     '@nuxtjs/onesignal',
-    [
-      // Doc: https://pwa.nuxtjs.org/
-      '@nuxtjs/pwa',
-      {
-        workbox: {
-          offlinePage: '/'
-          // Workbox options
-        },
-        meta: {
-          name: name
-        },
-        manifest: {
-          name: name,
-          short_name: 'Portfolio',
-          description: pkg.description,
-          lang: 'en'
-        },
-        icon: {
-          iconSrc: 'static/resources/images/icon/1x/primary-icon.png'
-          // Icon options
-        }
-      }
-    ],
-    [
-      'nuxt-i18n',
-      {
-        vueI18nLoader: true,
-        lazy: true,
-        langDir: './lang/',
-        baseUrl: baseUrl,
-        detectBrowserLanguage: {
-          useCookie: true,
-          cookieKey: 'kcnt-i18n'
-        },
-        locales: [
-          {
-            code: 'en',
-            name: 'english',
-            iso: 'en-US',
-            file: 'en.js'
-          },
-          {
-            code: 'th',
-            name: 'thai',
-            iso: 'th-TH',
-            file: 'th.js'
-          }
-        ],
-        defaultLocale: 'en'
-      }
-    ]
+    // Doc: https://pwa.nuxtjs.org/
+    ['@nuxtjs/pwa', require('./nuxt-config/pwa')(data)],
+    ['nuxt-i18n', require('./nuxt-config/i18n')(data)]
   ],
 
   /*
@@ -228,120 +117,24 @@ module.exports = {
   /*
    ** Vuetify module configuration
    */
-  vuetify: {
-    iconfont: 'fa4',
-    icons: {
-      moon: 'far fa-moon',
-      sun: 'far fa-sun',
-      version: 'fas fa-at',
-      home: 'fas fa-home',
-      beta: 'fas fa-flask',
-      admin: 'fas fa-user-shield',
-      github: 'fab fa-github',
-      docs: 'fas fa-book',
-      apis: 'fas fa-code'
-    },
-    treeShake: isProd
-    // Vuetify options
-    //  theme: { }
-  },
+  vuetify: require('./nuxt-config/vuetify')(data),
 
   /*
    ** Purge CSS module configuration (NO WORK, IT'S REMOVE USED CLASS)
    */
-  purgeCSS: {
-    // See https://github.com/Developmint/nuxt-purgecss
-    enabled: false,
-    mode: 'postcss',
-    paths: [
-      './components/**/*.vue',
-      './layouts/**/*.vue',
-      './pages/**/*.vue',
-      './plugins/**/*.js'
-    ],
-    styleExtensions: ['.css', '.scss'],
-    keyframes: true,
-    fontFace: true
-  },
+  purgeCSS: require('./nuxt-config/purgecss')(data),
 
   /*
    ** Sentry module configuration
    */
-  sentry: {
-    disabled: isDev,
-    public_key: 'ae4134e4a62b4ccd8bc0b7b7aab7e7c7',
-    project_id: '1338780',
-    config: {
-      // Additional config
-      environment: env,
-      release: isProd ? `portfolio@${pkg.version}` : undefined,
-      debug: isDev
-    }
-  },
+  sentry: require('./nuxt-config/sentry')(data),
 
-  oneSignal: {
-    cdn: true,
-    init: {
-      appId: branch === 'master' ? onesignal_prod : onesignal_dev,
-      autoRegister: true,
-      persistNotification: false, // dismiss the notification after 20 seconds
-      notifyButton: {
-        enable: false,
-        size: 'small',
-        displayPredicate: function() {
-          return OneSignal.isPushNotificationsEnabled().then(function(
-            isPushEnabled
-          ) {
-            /* The user is subscribed, so we want to return "false" to hide the Subscription Bell */
-            return !isPushEnabled
-          })
-        }
-      },
-      allowLocalhostAsSecureOrigin: true,
-      welcomeNotification: {
-        title: 'Welcome to portfolio website',
-        message: 'Thanks for subscribing!'
-      }
-    }
-  },
+  oneSignal: require('./nuxt-config/onesignal')(data),
 
-  redirect: [
-    // Redirect options here
-    {
-      from: '^/images/(.*)$',
-      to: '/resources/images/$1'
-    },
-    ...generateIconPath()
-  ],
+  redirect: require('./nuxt-config/redirect')(data),
 
   /*
    ** Build configuration
    */
-  build: {
-    extractCSS: true,
-    publicPath: '/_kcnt/',
-    splitChunks: {
-      layouts: true,
-      pages: true
-    },
-    analyze: {
-      analyzerMode: isDev ? 'static' : 'disable',
-      reportFilename: 'report.html',
-      openAnalyzer: false
-    },
-    /*
-     ** You can extend webpack config here
-     */
-    extend(config, ctx) {
-      // Run ESLint on save
-      if (ctx.isDev && ctx.isClient) {
-        config.module.rules.push({
-          enforce: 'pre',
-          test: /\.(js|vue)$/,
-          loader: 'eslint-loader',
-          exclude: /(node_modules)/
-        })
-      }
-    }
-  }
+  build: require('./nuxt-config/build')(data)
 }
